cloud_functions/budget_policy/requirements.txt
# Function dependencies, for example:
# package>=version 
openai>=0.27.6
google-cloud-storage>=1.30.0